♻️ Refactor: typeColorList 코드 분리
