+++ src/components/Detail.jsx
@@ -4,27 +4,7 @@ import HeartIcon from "../images/heart.svg?react";
 import HeartFillIcon from "../images/heart_fill.svg?react";
 import { add, remove } from "../RTK/slice";
 import { useEffect, useState } from "react";
-
-const typeColorList = {
-  bug: "#9F9F28",
-  dark: "#4F4747",
-  dragon: "#576FBC",
-  electric: "#DFBC28",
-  fairy: "#E18CE1",
-  fighting: "#E49021",
-  fire: "#E4613E",
-  flying: "#74AAD0",
-  ghost: "#6F4570",
-  grass: "#439837",
-  ground: "#A4733C",
-  ice: "#47C8C8",
-  normal: "#828282",
-  poison: "#9354CB",
-  psychic: "#E96C8C",
-  rock: "#A9A481",
-  steel: "#74B0CB",
-  water: "#3099E1",
-};
+import typeColorList from "../constants/typeColorList";
 
 const Detail = () => {
   const selectedId = useSelector((state) => state.selectedPokemon?.id);
